Add Day-3 Codeforces Problems Solutions
Getting harder now. A problem takes me 40 minutes+.
"Presents" took me probably more than an hour.
I had a hard time understanding the "Presents" problem,
so I ask ChatGPT to help me understand it. It turns out
the problem is asking who gave the present to the
number input. So for example the input is 2 3 4 1,
friend 1 gave to friend 2, friend 2 gave to friend 3,
friend 3 gave to friend 4, and friend 4 gave to friend 1.

Thought process:
Who gave friend 1 the gift that he gave to friend 2? It is
friend 4
Who gave friend 2 the gift the he gave to friend 3? It is
friend 1
Who gave friend 3 the gift that he gave to friend 4? It is
friend 2
Who gave friend 4 the gift that he gave to friend 1? It is
friend 3

So the output is 4 1 2 3.

Took me an effort to understand.

Signed-off-by: Christopher A. Quinto <cquinto.primary@gmail.com>